planner: split aggregation pushing into two files
